refactor: move state component classes to self files
Refs: SITKO-CORE-T-10
